update hft support and docs
